refactor: 위치 관련 유틸리티 함수 location 디렉토리로 정리
