+++ src/domains/Home/hooks/useWatchLocation.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Location } from '@/types';
-import geolocationErrorMessage from '@/utils/geolocationErrorMessage.ts';
-import toLocation from '@/utils/toLocation.ts';
+import geolocationErrorMessage from '@/utils/location/geolocationErrorMessage.ts';
+import toLocation from '@/utils/location/toLocation.ts';
 
 const useWatchLocation = () => {
   const [location, setLocation] = useState<Location>();
